performanceCount v3

+++ design-parent/src/main/java/org/bytedancer/crayzer/design_mode_principle/inaciton/performanceCounter/v1/Aggregator.java
@@ -31,7 +31,7 @@ public class Aggregator {
         if (count != 0) {
             avgRespTime = sumRespTime / count;
         }
-        long tps = (long) (count / durationInMillis * 1000);
+        long tps = count / durationInMillis * 1000;
         Collections.sort(requestInfos, new Comparator<RequestInfo>() {
             @Override
             public int compare(RequestInfo o1, RequestInfo o2) {
